+++ packages/graph-explorer/src/connector/gremlin/templates/neighborsCountTemplate.test.ts
@@ -1,25 +1,25 @@
 import neighborsCountTemplate from "./neighborsCountTemplate";
 
 describe("Gremlin > neighborsCountTemplate", () => {
-  it("Should return a template for the given vertex id with default limit", () => {
+  it("Should return a template for the given vertex id", () => {
     const template = neighborsCountTemplate({
       vertexId: "12",
       idType: "string",
     });
 
     expect(template).toBe(
-      'g.V("12").both().limit(500).dedup().group().by(label).by(count())'
+      'g.V("12").both().dedup().group().by(label).by(count())'
     );
   });
 
-  it("Should return a template for the given vertex id with default limit and number type", () => {
+  it("Should return a template for the given vertex id with number type", () => {
     const template = neighborsCountTemplate({
       vertexId: "12",
       idType: "number",
     });
 
     expect(template).toBe(
-      "g.V(12L).both().limit(500).dedup().group().by(label).by(count())"
+      "g.V(12L).both().dedup().group().by(label).by(count())"
     );
   });