+++ packages/graph-explorer/src/connector/testUtils/globalMockFetch.ts
@@ -8,18 +8,26 @@ const RESPONSES_FILES_MAP: Record<string, string> = {
   "7062d2e": `${GREMLIN}/edges-labels-and-counts.json`,
   "35be2501": `${GREMLIN}/should-return-1-random-node.json`,
   "54fa1494": `${GREMLIN}/should-return-airports-whose-code-matches-with-SFA.json`,
-  "77c63d2": `${GREMLIN}/should-return-all-neighbors-from-node-2018.json`,
+  e9e93b2: `${GREMLIN}/should-return-all-neighbors-from-node-2018.json`,
   "37e14b1": `${GREMLIN}/should-return-all-neighbors-from-node-2018-counts.json`,
-  "308b4988": `${GREMLIN}/should-return-filtered-neighbors-from-node-2018.json`,
+  "5698a21a": `${GREMLIN}/should-return-filtered-neighbors-from-node-2018.json`,
   "40a4690b": `${GREMLIN}/should-return-filtered-neighbors-from-node-2018-counts.json`,
-  "1614f686": `${GREMLIN}/should-return-neighbors-counts-for-node-123.json`,
+  "59bc2d43": `${GREMLIN}/should-return-neighbors-counts-for-node-123.json`,
 };
 
 const globalMockFetch = () => {
   // eslint-disable-next-line @typescript-eslint/ban-ts-comment
   // @ts-ignore
   global.fetch = jest.fn(async (url: string) => {
-    const response = await import(RESPONSES_FILES_MAP[shortHash(url)]);
+    const key = shortHash(url);
+    const filePath = RESPONSES_FILES_MAP[key];
+    if (!filePath) {
+      throw new Error(
+        `Failed to find a response file in the map for key '${key}'`,
+        { cause: { url } }
+      );
+    }
+    const response = await import(filePath);
     return Promise.resolve({
       json: () => {
         return Promise.resolve(response);